+++ src/chebyshev/mod.rs
@@ -6,8 +6,12 @@
 //!
 //! This rule can integrate formulas on the form f(x) * (1 - x^2)^`a` on finite intervals, where `a` is either -1/2 or 1/2.
 
-use crate::{__impl_node_weight_rule, Node, Weight};
+use crate::{
+    __impl_node_weight_rule, Node, Weight,
+    math::{cos, sin},
+};
 
+use alloc::boxed::Box;
 use core::{f64::consts::PI, num::NonZeroUsize};
 
 #[cfg(feature = "rayon")]
@@ -44,7 +48,7 @@ impl GaussChebyshevFirstKind {
         Self {
             node_weight_pairs: (1..degree.get() + 1)
                 .rev()
-                .map(|i| ((PI * (2.0 * (i as f64) - 1.0) / (2.0 * n)).cos(), PI / n))
+                .map(|i| (cos(PI * (2.0 * (i as f64) - 1.0) / (2.0 * n)), PI / n))
                 .collect(),
         }
     }
@@ -66,7 +70,7 @@ impl GaussChebyshevFirstKind {
             node_weight_pairs: (1..degree.get() + 1)
                 .into_par_iter()
                 .rev()
-                .map(|i| ((PI * (2.0 * (i as f64) - 1.0) / (2.0 * n)).cos(), PI / n))
+                .map(|i| (cos(PI * (2.0 * (i as f64) - 1.0) / (2.0 * n)), PI / n))
                 .collect(),
         }
     }
@@ -143,9 +147,9 @@ impl GaussChebyshevSecondKind {
                 .rev()
                 .map(|i| {
                     let over_n_plus_1 = 1.0 / (n + 1.0);
-                    let sin_val = (PI * i as f64 * over_n_plus_1).sin();
+                    let sin_val = sin(PI * i as f64 * over_n_plus_1);
                     (
-                        (PI * i as f64 * over_n_plus_1).cos(),
+                        cos(PI * i as f64 * over_n_plus_1),
                         PI * over_n_plus_1 * sin_val * sin_val,
                     )
                 })
@@ -164,9 +168,9 @@ impl GaussChebyshevSecondKind {
                 .rev()
                 .map(|i| {
                     let over_n_plus_1 = 1.0 / (n + 1.0);
-                    let sin_val = (PI * i as f64 * over_n_plus_1).sin();
+                    let sin_val = sin(PI * i as f64 * over_n_plus_1);
                     (
-                        (PI * i as f64 * over_n_plus_1).cos(),
+                        cos(PI * i as f64 * over_n_plus_1),
                         PI * over_n_plus_1 * sin_val * sin_val,
                     )
                 })
@@ -225,6 +229,7 @@ __impl_node_weight_rule! {GaussChebyshevSecondKind, GaussChebyshevSecondKindNode
 
 #[cfg(test)]
 mod test {
+    use crate::math::{cos, sin};
     use approx::assert_abs_diff_eq;
 
     use super::{GaussChebyshevFirstKind, GaussChebyshevSecondKind};
@@ -308,8 +313,8 @@ mod test {
         for (i, (x, w)) in rule.into_iter().enumerate() {
             // Source: https://en.wikipedia.org/wiki/Chebyshev%E2%80%93Gauss_quadrature
             let ii = deg - i as f64;
-            let x_should = (ii * PI / (deg + 1.0)).cos();
-            let w_should = PI / (deg + 1.0) * (ii * PI / (deg + 1.0)).sin().powi(2);
+            let x_should = cos(ii * PI / (deg + 1.0));
+            let w_should = PI / (deg + 1.0) * sin(ii * PI / (deg + 1.0)).powi(2);
 
             assert_abs_diff_eq!(x, x_should);
             assert_abs_diff_eq!(w, w_should);