+++ src/hermite/mod.rs
@@ -26,8 +26,9 @@
 #[cfg(feature = "rayon")]
 use rayon::prelude::{IntoParallelRefIterator, ParallelIterator};
 
-use crate::{__impl_node_weight_rule, DMatrixf64, Node, Weight};
+use crate::{__impl_node_weight_rule, DMatrixf64, Node, Weight, math::sqrt};
 
+use alloc::boxed::Box;
 use core::{f64::consts::PI, num::NonZeroUsize};
 
 /// A Gauss-Hermite quadrature scheme.
@@ -71,10 +72,11 @@ impl GaussHermite {
     /// see Gil, Segura, Temme - Numerical Methods for Special Functions
     pub fn new(deg: NonZeroUsize) -> Self {
         let mut companion_matrix = DMatrixf64::from_element(deg.get(), deg.get(), 0.0);
+
         // Initialize symmetric companion matrix
         for idx in 0..deg.get() - 1 {
             let idx_f64 = 1.0 + idx as f64;
-            let element = (idx_f64 * 0.5).sqrt();
+            let element = sqrt(idx_f64 * 0.5);
             companion_matrix[(idx, idx + 1)] = element;
             companion_matrix[(idx + 1, idx)] = element;
         }
@@ -90,7 +92,7 @@ impl GaussHermite {
                 eigen
                     .eigenvectors
                     .row(0)
-                    .map(|x| x * x * PI.sqrt())
+                    .map(|x| x * x * sqrt(PI))
                     .iter()
                     .copied(),
             )
@@ -150,7 +152,7 @@ mod tests {
     #[test]
     fn check_degree_1() {
         let rule = GaussHermite::new(1.try_into().unwrap());
-        assert_eq!(rule.as_node_weight_pairs(), &[(0.0, PI.sqrt())]);
+        assert_eq!(rule.as_node_weight_pairs(), &[(0.0, sqrt(PI))]);
         for constant in (1..100).step_by(10) {
             assert_abs_diff_eq!(rule.integrate(|x| f64::from(constant) * x), 0.0);
         }
@@ -185,7 +187,7 @@ mod tests {
     #[test]
     fn check_iterators() {
         let rule = GaussHermite::new(3.try_into().unwrap());
-        let ans = core::f64::consts::PI.sqrt() / 2.0;
+        let ans = sqrt(PI) / 2.0;
 
         assert_abs_diff_eq!(
             ans,
@@ -212,7 +214,7 @@ mod tests {
     fn integrate_one() {
         let quad = GaussHermite::new(5.try_into().unwrap());
         let integral = quad.integrate(|_x| 1.0);
-        assert_abs_diff_eq!(integral, PI.sqrt(), epsilon = 1e-14);
+        assert_abs_diff_eq!(integral, sqrt(PI), epsilon = 1e-14);
     }
 
     #[cfg(feature = "rayon")]
@@ -220,6 +222,6 @@ mod tests {
     fn par_integrate_one() {
         let quad = GaussHermite::new(5.try_into().unwrap());
         let integral = quad.par_integrate(|_x| 1.0);
-        assert_abs_diff_eq!(integral, PI.sqrt(), epsilon = 1e-15);
+        assert_abs_diff_eq!(integral, sqrt(PI), epsilon = 1e-15);
     }
 }