+++ src/laguerre/mod.rs
@@ -24,9 +24,12 @@
 #[cfg(feature = "rayon")]
 use rayon::prelude::{IntoParallelRefIterator, ParallelIterator};
 
-use crate::gamma::gamma;
-use crate::{__impl_node_weight_rule, DMatrixf64, FiniteAboveNegOneF64, Node, Weight};
+use crate::{
+    __impl_node_weight_rule, DMatrixf64, FiniteAboveNegOneF64, Node, Weight,
+    math::{gamma, sqrt},
+};
 
+use alloc::boxed::Box;
 use core::num::NonZeroUsize;
 
 /// A Gauss-Laguerre quadrature scheme.
@@ -76,7 +79,7 @@ impl GaussLaguerre {
         // Initialize symmetric companion matrix
         for idx in 0..degree.get() - 1 {
             let idx_f64 = 1.0 + idx as f64;
-            let off_diag = (idx_f64 * (idx_f64 + alpha.get())).sqrt();
+            let off_diag = sqrt(idx_f64 * (idx_f64 + alpha.get()));
             companion_matrix[(idx, idx)] = diag;
             companion_matrix[(idx, idx + 1)] = off_diag;
             companion_matrix[(idx + 1, idx)] = off_diag;